moved code from loglike.cc to library
